refactor: moved UserState.java
