+++ bot/src/main/java/edu/java/bot/commands/UntrackCommand.java
@@ -2,7 +2,7 @@ package edu.java.bot.commands;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-import edu.java.bot.model.UserState;
+import edu.java.bot.repository.in_memory.UserState;
 import org.springframework.stereotype.Component;
 import static edu.java.bot.repository.in_memory.DBUsersLinks.deleteLink;
 import static edu.java.bot.repository.in_memory.DBUsersLinks.isUserHasLink;
