+++ bot/src/main/java/edu/java/bot/repository/in_memory/UserState.java
@@ -1,4 +1,4 @@
-package edu.java.bot.model;
+package edu.java.bot.repository.in_memory;
 
 public enum UserState {
     UNREGISTERED, REGISTERED, TRACK, UNTRACKED
