+++ bot/src/test/java/UserStateChangeTest.java
@@ -14,10 +14,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import static edu.java.bot.commands.TrackCommand.trackURL;
 import static edu.java.bot.commands.UntrackCommand.untrackURL;
-import static edu.java.bot.model.UserState.REGISTERED;
-import static edu.java.bot.model.UserState.TRACK;
-import static edu.java.bot.model.UserState.UNREGISTERED;
-import static edu.java.bot.model.UserState.UNTRACKED;
+import static edu.java.bot.repository.in_memory.UserState.REGISTERED;
+import static edu.java.bot.repository.in_memory.UserState.TRACK;
+import static edu.java.bot.repository.in_memory.UserState.UNREGISTERED;
+import static edu.java.bot.repository.in_memory.UserState.UNTRACKED;
 import static edu.java.bot.repository.in_memory.DBUsersState.getUserState;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
